# docs/openapi/admin_openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AFF Admin API",
    "version": "0.1.0",
    "description": "Admin Courier account management and read-only Delivery oversight (stories E1 and E11). Design intent lives in docs/api_design.md §11; this document reflects the routes as implemented in backend/src/modules/admin. The remaining §11 routes — GET /admin/users (G1), PATCH /admin/users/{id}/status (G2), GET /admin/listings (G4) and PATCH /admin/listings/{id}/cancel (G3) — are registered but still return 501, and are deliberately not documented here until they are wired."
  },
  "servers": [
    { "url": "http://localhost:5000/api", "description": "Local development" }
  ],
  "tags": [
    { "name": "Admin Couriers", "description": "Courier account creation and roster" },
    { "name": "Admin Deliveries", "description": "Read-only oversight of the Courier pipeline" }
  ],
  "paths": {
    "/admin/couriers": {
      "post": {
        "tags": ["Admin Couriers"],
        "summary": "Create a Courier account",
        "description": "The only way a Courier account is ever created — neither auth registration endpoint accepts role=COURIER (E1). Creates the USER row (role=COURIER) and its COURIER profile together; if the profile write fails, the just-created USER is deleted so the email is not permanently blocked. tempPassword is set by the Admin at creation time, is held to the same strength rules as any registration password, and is an ordinary password from then on — there is no forced-reset flow behind it. The new account is immediately visible in GET /admin/couriers.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateCourierRequest" },
              "example": {
                "username": "courier_01",
                "email": "courier01@aff.com",
                "tempPassword": "Str0ng#Pass",
                "fullName": "Nguyen Van A"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created Courier account",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "data": { "$ref": "#/components/schemas/Courier" } },
                  "required": ["data"]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      },
      "get": {
        "tags": ["Admin Couriers"],
        "summary": "List Courier accounts",
        "description": "Paginated Courier roster, newest account first. Couriers carry the same account-management fields as Recipients and Donors (id, username, email, role, status), plus the Courier's fullName. Also reachable, once G1 is implemented, via GET /admin/users?role=COURIER — this route is the convenience alias the PRD calls out separately.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/Limit" }
        ],
        "responses": {
          "200": {
            "description": "One page of Courier accounts",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/PaginatedEnvelope" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "items": {
                              "type": "array",
                              "items": { "$ref": "#/components/schemas/Courier" }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/admin/deliveries": {
      "get": {
        "tags": ["Admin Deliveries"],
        "summary": "List every Delivery (read-only)",
        "description": "Paginated oversight of the whole Courier pipeline, newest first, optionally narrowed to one stage. Each row carries the assigned Courier (null while unclaimed), the stage, the lifecycle timestamps, and the associated Order's recipient reference.\n\nRead-only by design: there is no endpoint anywhere in this API to assign, reassign, or force-claim a Delivery to a Courier. Admin manual dispatch is an explicit scope boundary (E11).\n\npickupAddressText and pickupAddressLocation are omitted from these rows — the oversight table shows pipeline state, not navigation detail, so the Donor address is not resolved per row. A Courier working a Delivery gets those fields from GET /deliveries/active and the claim/pickup/deliver responses instead (§9).",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/Limit" },
          {
            "name": "stage",
            "in": "query",
            "required": false,
            "description": "Narrow the table to one Delivery stage. Omit to see every stage.",
            "schema": { "$ref": "#/components/schemas/DeliveryStage" }
          }
        ],
        "responses": {
          "200": {
            "description": "One page of Deliveries",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/PaginatedEnvelope" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "items": {
                              "type": "array",
                              "items": { "$ref": "#/components/schemas/AdminDelivery" }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "parameters": {
      "Page": {
        "name": "page",
        "in": "query",
        "required": false,
        "description": "1-based page number.",
        "schema": { "type": "integer", "minimum": 1, "default": 1 }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "Rows per page.",
        "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": { "message": { "type": "string" } },
        "required": ["message"]
      },
      "PaginatedEnvelope": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "items": { "type": "array", "items": {} },
              "page": { "type": "integer" },
              "limit": { "type": "integer" },
              "total": { "type": "integer" }
            },
            "required": ["items", "page", "limit", "total"]
          }
        },
        "required": ["data"]
      },
      "DeliveryStage": {
        "type": "string",
        "enum": ["AWAITING_COURIER", "ASSIGNED", "PICKED_UP", "DELIVERED", "CANCELLED"]
      },
      "GeoLocation": {
        "type": "object",
        "properties": {
          "latitude": { "type": "number" },
          "longitude": { "type": "number" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "CreateCourierRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "username": {
            "type": "string",
            "description": "English letters, numbers, underscores and hyphens only."
          },
          "email": { "type": "string", "description": "Normalised to lowercase." },
          "tempPassword": {
            "type": "string",
            "minLength": 8,
            "maxLength": 72,
            "description": "Same strength rules as any registration password: at least 8 characters, one number, one capital, and one special character."
          },
          "fullName": { "type": "string", "minLength": 1 }
        },
        "required": ["username", "email", "tempPassword", "fullName"]
      },
      "Courier": {
        "type": "object",
        "description": "A Courier account: the shared account-management fields plus the Courier's name.",
        "properties": {
          "id": { "type": "string" },
          "username": { "type": "string" },
          "email": { "type": "string" },
          "role": { "type": "string", "enum": ["COURIER"] },
          "country": { "type": "string", "nullable": true },
          "city": { "type": "string", "nullable": true },
          "status": { "type": "string", "enum": ["ACTIVE", "DEACTIVATED"] },
          "avatarUrl": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "fullName": {
            "type": "string",
            "description": "Empty string if the COURIER profile row is missing — the account is still listed so a half-written account stays visible to the Admin."
          }
        },
        "required": ["id", "username", "email", "role", "status", "createdAt", "fullName"]
      },
      "AdminDelivery": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "orderId": { "type": "string" },
          "courierId": { "type": "string", "nullable": true },
          "stage": { "$ref": "#/components/schemas/DeliveryStage" },
          "pickupAddressText": {
            "type": "string",
            "nullable": true,
            "description": "Always absent on this endpoint — see the route description."
          },
          "pickupAddressLocation": {
            "allOf": [{ "$ref": "#/components/schemas/GeoLocation" }],
            "nullable": true,
            "description": "Always absent on this endpoint — see the route description."
          },
          "pickedUpAt": { "type": "string", "format": "date-time", "nullable": true },
          "deliveredAt": { "type": "string", "format": "date-time", "nullable": true },
          "courierLastLocation": {
            "allOf": [{ "$ref": "#/components/schemas/GeoLocation" }],
            "nullable": true,
            "description": "Only populated while stage=PICKED_UP."
          },
          "createdAt": { "type": "string", "format": "date-time" },
          "deliveryAddressText": {
            "type": "string",
            "nullable": true,
            "description": "Where the order is being delivered to — the Recipient's address."
          },
          "deliveryLocation": {
            "allOf": [{ "$ref": "#/components/schemas/GeoLocation" }],
            "nullable": true,
            "description": "Coordinates of the delivery destination. The Courier's map re-centres here after pickup."
          },
          "requiresCashCollection": {
            "type": "boolean",
            "description": "Whether the Courier must collect cash on delivery. Derived server-side from the Order's payment method, which is never itself exposed — a Courier needs to know whether to collect money, not how the Recipient paid."
          },
          "courier": {
            "type": "object",
            "nullable": true,
            "description": "Null while the Delivery is unclaimed.",
            "properties": {
              "id": { "type": "string" },
              "fullName": { "type": "string" }
            },
            "required": ["id", "fullName"]
          },
          "order": {
            "type": "object",
            "properties": {
              "id": { "type": "string" },
              "recipientId": {
                "type": "string",
                "nullable": true,
                "description": "Null only as a data-integrity fallback — every DELIVERY is created from an ORDER, so in practice this is always set. Degrading one row is preferred to failing the whole oversight table."
              }
            },
            "required": ["id", "recipientId"]
          }
        },
        "required": ["id", "orderId", "courierId", "stage", "createdAt", "courier", "order"]
      }
    },
    "responses": {
      "ValidationError": {
        "description": "A field failed validation, an unknown key was sent, or the stage filter is not a valid Delivery stage",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "examples": {
              "weakPassword": {
                "summary": "tempPassword fails the shared strength rules",
                "value": { "message": "Password must be at least 8 characters." }
              },
              "missingFullName": {
                "summary": "fullName omitted or blank",
                "value": { "message": "Full name is required." }
              },
              "badStage": {
                "summary": "stage filter outside the Delivery lifecycle",
                "value": { "message": "Stage must be a valid Delivery stage." }
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing, malformed, expired, or already-revoked token",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "message": "Authentication is required." }
          }
        }
      },
      "Forbidden": {
        "description": "The caller is authenticated but is not an Admin",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "message": "You do not have permission to access this resource." }
          }
        }
      },
      "Conflict": {
        "description": "The submitted email is already registered to another account",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "message": "This email is already registered." }
          }
        }
      }
    }
  }
}